requirements.txt
selenium==4.27.1
python-dotenv==1.0.1
fastapi[standard]==0.115.0
webdriver-manager==4.0.2
pyjwt==2.9.0
starlette==0.38.6
newrelic==10.2.0
cryptography==44.0.0